feat: :sparkles: add card point
